refector: move m.u.q to a differrent file
